+++ experiments/calibration/models/snowy_land.jl
@@ -45,7 +45,7 @@ function setup_model(
         start_date,
         stop_date,
     )
-
+    prognostic_land_components = (:canopy, :snow, :soil)
     # Snow model setup
     # Set β = 0 in order to regain model without density dependence
     α_snow = Snow.ZenithAngleAlbedoModel(toml_dict)
@@ -58,7 +58,7 @@ function setup_model(
         forcing,
         toml_dict,
         Δt;
-        prognostic_land_components = (:canopy, :snow, :soil),
+        prognostic_land_components,
         α_snow,
         scf,
     )
@@ -69,18 +69,38 @@ function setup_model(
     conductance = PModelConductance{FT}(toml_dict)
     soil_moisture_stress =
         ClimaLand.Canopy.PiecewiseMoistureStressModel{FT}(domain, toml_dict)
+    biomass = ClimaLand.Canopy.PrescribedBiomassModel{FT}(
+        domain,
+        LAI,
+        toml_dict;
+        height = ClimaLand.Canopy.clm_canopy_height(
+            surface_space;
+            max_height = atmos.h * FT(0.9),
+        ),
+    )
     canopy = ClimaLand.Canopy.CanopyModel{FT}(
         surface_domain,
         canopy_forcing,
         LAI,
         toml_dict;
-        prognostic_land_components = (:canopy, :snow, :soil),
+        prognostic_land_components,
         photosynthesis,
         conductance,
         soil_moisture_stress,
+        biomass,
     )
+
     # Construct the land model with all default components except for snow
-    land = LandModel{FT}(forcing, LAI, toml_dict, domain, Δt; snow, canopy)
+    land = LandModel{FT}(
+        forcing,
+        LAI,
+        toml_dict,
+        domain,
+        Δt;
+        snow,
+        canopy,
+        prognostic_land_components,
+    )
     return land
 end