+++ NEWS.md
@@ -3,6 +3,8 @@ ClimaLand.jl Release Notes
 
 main
 ------
+- Updated parameters, quantum yield for C4, canopy LW PR[#1613](https://github.com/CliMA/ClimaLand.jl/pull/1613)
+
 v1.5.2
 ------
 - Fix divide by zero bug
@@ -10,6 +12,7 @@ v1.5.2
 v1.5.1
 ------
 - Adjust RootSolvers lower compat 
+
 v1.5.0
 ------
 - Update to Insolation v1; use cosine zenith angle directly from Insolation.jl PR[#1617](https://github.com/CliMA/ClimaLand.jl/pull/1617)