+++ NEWS.md
@@ -6,6 +6,7 @@ main
 
 v1.5.3
 ------
+- Updated parameters, quantum yield for C4, canopy LW PR[#1613](https://github.com/CliMA/ClimaLand.jl/pull/1613)
 
 v1.5.2
 ------
@@ -13,7 +14,8 @@ v1.5.2
 
 v1.5.1
 ------
-- Adjust RootSolvers lower compat
+- Adjust RootSolvers lower compat 
+
 v1.5.0
 ------
 - Update to Insolation v1; use cosine zenith angle directly from Insolation.jl PR[#1617](https://github.com/CliMA/ClimaLand.jl/pull/1617)